pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "macos-universal-updator"
version = "2.0.0"
description = "A beautiful TUI application for managing macOS system and application updates"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "FuadPoroshtica", email = "fuad@poroshtica.com"}
]
keywords = ["macos", "updater", "tui", "homebrew", "system-update"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS :: MacOS X",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Systems Administration",
    "Topic :: Utilities",
]
dependencies = [
    "textual>=0.47.0",
    "rich>=13.7.0",
    "pyyaml>=6.0.1",
    "psutil>=5.9.0",
    "packaging>=23.0",
    "click>=8.1.0",
    "apscheduler>=3.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
macos-updater = "updater.cli:main"
updater = "updater.cli:main"

[project.urls]
Homepage = "https://github.com/FuadPoroshtica/MacOS-Universal-Updator"
Repository = "https://github.com/FuadPoroshtica/MacOS-Universal-Updator"
Issues = "https://github.com/FuadPoroshtica/MacOS-Universal-Updator/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
